read_attr(): UNIT bugfix for dsName with date info (#630)
+ utils.readfile.read_attribute(): fix a bug of grabbing the UNIT for HDF5 dataset, if `datasetName` is specified with date info as suffix. This only affects the unit display in the colorbar for plotting.
+ utils.utils1.get_residual_std/rms(): use np.float32 to replace np.float to avoid numpy deprecation warning
+ view: print WARNING msg regardless of --verbose setting
+ tsview: more friendly figsize_img API for the interactive tsview.ipynb in MintPy-tutorial repo
+ plot_network: fix --nodisplay bug while calling from jupyter notebook
+ install: use bash *.sh to replace chmod +x cmd in docs/installation.md and .circleci/config.yml files

+++ mintpy/utils/utils1.py
@@ -84,7 +84,7 @@ def get_residual_std(timeseries_resid_file, mask_file='maskTempCoh.h5', ramp_typ
     # Read residual std text file
     print('read timeseries RMS from file: '+std_file)
     fc = np.loadtxt(std_file, dtype=bytes).astype(str)
-    std_list = fc[:, 1].astype(np.float).tolist()
+    std_list = fc[:, 1].astype(np.float32).tolist()
     date_list = list(fc[:, 0])
     return std_list, date_list
 
@@ -135,7 +135,7 @@ def get_residual_rms(timeseries_resid_file, mask_file='maskTempCoh.h5', ramp_typ
     # Read residual RMS text file
     print('read timeseries residual RMS from file: '+rms_file)
     fc = np.loadtxt(rms_file, dtype=bytes).astype(str)
-    rms_list = fc[:, 1].astype(np.float).tolist()
+    rms_list = fc[:, 1].astype(np.float32).tolist()
     date_list = list(fc[:, 0])
     return rms_list, date_list, rms_file
